backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amadeus-gateway"
version = "0.1.0"
description = "Backend services for the Amadeus 2.0 gateway"
readme = "../README.md"
requires-python = ">=3.10"
authors = [{ name = "Amadeus" }]
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<0.30.0",
    "structlog>=24.1.0,<25.0.0",
    "pydantic>=2.6.0,<3.0.0",
    "pydantic-settings>=2.2.0,<3.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0,<0.30.0",
    "alembic>=1.13.0,<2.0.0",
    "redis>=5.0.0,<6.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "email-validator>=2.1.0,<3.0.0",
    "argon2-cffi>=23.1.0,<24.0.0",
    "PyJWT>=2.8.0,<3.0.0",
    "pyotp>=2.9.0,<3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.1.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
]

[tool.ruff]
line-length = 100
src = ["backend"]
target-version = "py311"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E402", "E501", "E701", "E702"]

[tool.ruff.lint.per-file-ignores]
"backend/gateway/app/__init__.py" = ["F401"]
